+++ raiden-ts/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## [Unreleased]
+
+## [0.9.0] - 2020-05-28
+
 ### Added
 - [#1473] Expose config$ observable
 
@@ -197,7 +200,8 @@
 - Add protocol message implementation.
 
 
-[Unreleased]: https://github.com/raiden-network/light-client/compare/v0.8.0...HEAD
+[Unreleased]: https://github.com/raiden-network/light-client/compare/v0.9.0...HEAD
+[0.9.0]: https://github.com/raiden-network/light-client/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/raiden-network/light-client/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/raiden-network/light-client/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/raiden-network/light-client/compare/v0.5.2...v0.6.0